requirements.txt
numpy==2.0.2
python-dotenv==1.0.1
requests==2.32.3
pydub==0.25.1
